test: fix html tests
